完成了所有功能，做了数值优化，体验优化，本Demo开发结束

+++ Assets/Scripts/Data/TowerInfo.cs
@@ -8,7 +8,7 @@ public class TowerInfo
     public string name;
     public int money;
     public int atk;
-    public int atkRange;
+    public float atkRange;
     public float offsetTime;
     public int nextLev;
     public string imgRes;
